refactor: BaseTimeEntity.java를 domain 하위 common 패키지로 이동. (#19)
