+++ src/main/java/com/newfit/reservation/domain/User.java
@@ -1,5 +1,6 @@
 package com.newfit.reservation.domain;
 
+import com.newfit.reservation.domain.common.BaseTimeEntity;
 import com.newfit.reservation.domain.dev.Proposal;
 import com.newfit.reservation.domain.dev.Report;
 import jakarta.persistence.*;
